Move `GetSystemDateTime()` to StringConversion
This method may be more generally useful, particularly to other logging
code.
